% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ackley.R
\name{ackley}
\alias{ackley}
\title{Ackley function}
\usage{
ackley(x)
}
\arguments{
\item{x}{vector for calculating the Ackley function value.}
}
\value{
If \bold{x} is a vector, returns the Ackley function value corrisponding to this vector.
If is a matrix, returns
the value of the function applied on each row vector.
}
\description{
This function is a classic benchmark function for testing optimization algorithms.
}
\details{
Function has many local minima and represents a threat for almost any optimization algorithm.
The recommended values for the parameters are used, with a=20, b=0.2 and c=2*pi.
Usually x_i is set between [-32,32] for each i, but can also be restricted to a smaller domain.
}
\examples{
##INPUT:
xx = runif(3)
g <- ackley(xx)

}
