% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEbase.R, R/DEbudget.R
\name{DEbase}
\alias{DEbase}
\title{Base DE}
\usage{
DEbase(funcname, lo, up, n, NP, n_gen, f, CR, strategy = 1, ...)

DEbase(funcname, lo, up, n, NP, n_gen, f, CR, strategy = 1, ...)
}
\arguments{
\item{funcname}{function to optimize}

\item{lo}{lower bound (same for each component of the vector x)}

\item{up}{upper bound (same for each component of the vector x)}

\item{n}{problem size}

\item{NP}{population size (should be ~ four times the size of the problem)}

\item{n_gen}{number of generations (~ ten times the value of NP)}

\item{f}{F value (usually from 0.1 to 1.1)}

\item{CR}{crossover rate (it is a probability, should stay between 0 and 1)}

\item{strategy}{type of strategy implemented. Default to DE/RAND/1}

\item{...}{parameters to pass at objective function to optimize}
}
\value{
The output of the function \code{DEbase} is a list (of length 3) containing the following elements:\cr

\itemize{
\item \code{f_best}: the best value found by the algorithm for the last generation
\item \code{x_best}: the vector which corresponds to the best overall function value
\item \code{f}: the vector of the \code{n_gen} optimal values found by the algorithm at each generation
}

The output of the function \code{DEbase} is a list (of length 3) containing the following elements:\cr

\itemize{
\item \code{f_best}: the best value found by the algorithm for the last generation
\item \code{x_best}: the vector which corresponds to the best overall function value
\item \code{f}: the vector of the \code{n_gen} optimal values found by the algorithm at each generation
}
}
\description{
Differential Evolution procedure to solve minimization problems with bound constraints

Differential Evolution algorithm that deals with bound constraints and a budget constraint. Specific for portfolio optimization applications.
}
\examples{
##implementation of the schwefel function
schwef <- function(x)
{
 if(is.vector(x))
 {
   d <- length(x)
   sum <- sum(x*sin(sqrt(abs(x))))
   y <- 418.9829*d - sum
   return(y)
 }
 if(is.matrix(x))
 {
   d <- ncol(x)
   sum <- apply(x*sin(sqrt(abs(x))),1,sum)
   y <- 418.9829*d - sum
   return(y)
 }
}
##application of DEbase function
set.seed(123)
d <- DEbase(schwef,-500,500,10,40,400,0.8,0.4,strategy=2)

#examplezzzzz

}
\references{
S. Das, S. Mullick, P. N. Suganthan, \emph{Recent advances in differential evolution--
an updated survey}. Swarm and evolutionary computation, vol. 23, 2016, pp. 1--30

S. Das, S. Mullick, P. N. Suganthan, \emph{Recent advances in differential evolution--
an updated survey}. Swarm and evolutionary computation, vol. 23, 2016, pp. 1--30
Inserire reference per la procedura takkur
}
