% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEbudget.R
\name{DEbudget}
\alias{DEbudget}
\title{Budget DE}
\usage{
DEbudget(funcname, lo, up, n, NP, n_gen, f, CR, strategy, ...)
}
\arguments{
\item{funcname}{function to optimize}

\item{lo}{lower bound (same for each component of the vector x)}

\item{up}{upper bound (same for each component of the vector x)}

\item{n}{problem size}

\item{NP}{population size (should be ~ four times the size of the problem)}

\item{n_gen}{number of generations (~ ten times the value of NP)}

\item{f}{F value (usually from 0.1 to 1.1)}

\item{CR}{crossover rate (it is a probability, should stay between 0 and 1)}

\item{strategy}{type of strategy implemented.}

\item{...}{parameters to pass at objective function to optimize}
}
\value{
The output of the function \code{DEbase} is a list (of length 3) containing the following elements:\cr

\itemize{
\item \code{f_best}: the best value found by the algorithm for the last generation
\item \code{x_best}: the vector which corresponds to the best overall function value
\item \code{f}: the vector of the \code{n_gen} optimal values found by the algorithm at each generation
}
}
\description{
Differential Evolution algorithm that deals with bound constraints and a budget constraint. Specific for portfolio optimization applications.
}
\examples{
#examplezzzzz

}
\references{
S. Das, S. Mullick, P. N. Suganthan, \emph{Recent advances in differential evolution--
an updated survey}. Swarm and evolutionary computation, vol. 23, 2016, pp. 1--30
Inserire reference per la procedura takkur
}
