% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/griewank.R
\name{griewank}
\alias{griewank}
\title{Griewank function}
\usage{
griewank(x)
}
\arguments{
\item{x}{vector for calculating the Griewank function value.}
}
\value{
If \bold{x} is a vector, returns the Griewank function value corrisponding to this vector.
If is a matrix, returns
the value of the function applied on each row vector.
}
\description{
This function is a classic benchmark function for testing optimization algorithms.
}
\details{
Complex function with many local minima, which are regularly distributed. Usually evaluated
with x_i in [-600,600] for each i. Global minimum f(\bold{x'})=0 in \bold{x'}=(0,...,0).
}
\examples{

##INPUT:
xx = runif(3)
g <- griewank(xx)

}
