% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modsharpe.R
\name{ModSharpe}
\alias{ModSharpe}
\title{Modified Sharpe Ratio function}
\usage{
ModSharpe(x, mu, rf, Sigma)
}
\arguments{
\item{x}{a vector or matrix - if it is a matrix, the function will return the value of the ModSharpe for each row.}

\item{mu}{vector of the expected returns of the n risky assets}

\item{rf}{risk-free rate}

\item{Sigma}{estimator of the covariance matrix}
}
\value{
ModSharpe() returns as object the value of the function modified sharpe ratio calculated in x. If x is a matrix,
it returns a column vector and calculates the modified sharpe for each row.
}
\description{
Function that calculates the modified version of the Sharpe ratio, more coherent
when the expected excess portfolio return is negative.
}
\examples{

r <- matrix( ,nrow =60 ,ncol=25)
for (i in 1:60) {
 r[i,] <- runif(25,-2.5,2.5)
}
mu <- rowMeans(t(r))
Sigma <- cov(r)
xx <- runif(25,0.01,0.30)
ModSharpe(xx,mu,0,Sigma)

}
