% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schwefel.R
\name{schwefel}
\alias{schwefel}
\title{Schwefel function}
\usage{
schwefel(x)
}
\arguments{
\item{x}{vector for calculating the schwefel function value.}
}
\value{
If \bold{x} is a vector, returns the schwefel function value corrisponding to this vector.
If is a matrix, returns
the value of the function applied on each row vector.
}
\description{
This function is a classic benchmark function for testing optimization algorithms.
}
\details{
This function is complex with many local minima. The function presents a global minimum at
\bold{x'} = (420.9687,...,420.9687) and f(\bold{x'})=0.
Usually is evaluated on the hypercube denoted by: x_i in [-500,500].
}
\examples{
##INPUT:
xx = runif(3)
s <- schwefel(xx)

}
