% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{sphere}
\alias{sphere}
\title{Sphere function}
\usage{
sphere(x)
}
\arguments{
\item{x}{vector for calculating the sphere function value.}
}
\value{
If \bold{x} is a vector, returns the sphere function value corrisponding to this vector.
If is a matrix, returns
the value of the function applied on each row vector.
}
\description{
This function is a classic benchmark function for testing optimization algorithms.
}
\details{
The function presents a global minimum at
\bold{x'} = (0,...,0) and f(\bold{x'})=0.
Usually is evaluated on the hypercube denoted by: x_i in  [-5.12,5.12] for each i.
}
\examples{
##INPUT:
xx = runif(3)
s <- sphere(xx)

}
